backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# MLX
mlx>=0.18.0
mlx-lm

# SSE 流式响应
sse-starlette>=2.0.0

# 数据库
aiosqlite>=0.19.0

# 数据模型
pydantic>=2.5.0
pydantic-settings>=2.0.0

# 测试依赖
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0
